Validacioes

+++ src/app/registrar/registrar.component.ts
@@ -1,5 +1,7 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
+import Swal from 'sweetalert2';
+
 
 @Component({
   selector: 'app-registrar',
@@ -17,22 +19,58 @@ export class RegistrarComponent {
   };
 
   guardarDatosLocalStorage() {
-    // Agregar el nuevo usuario a la lista de usuarios registrados
-    this.usuariosRegistrados.push(this.users);
+    let emailRegex = /^[-\w.%+]{1,64}@(?:[A-Z0-9-]{1,63}\.){1,125}[A-Z]{2,63}$/i;
+    if(this.users.email=='' || this.users.username=='' || this.users.password==''){
+      Swal.fire({
+        icon: "error",
+        title: "Oops...",
+        text: "Complete Todos los Campos!",
+      });
+    }
+    else{
 
-    // Guardar la lista actualizada en el local storage
-    localStorage.setItem('users', JSON.stringify(this.usuariosRegistrados));
+    if(emailRegex.test(this.users.email)){
+      if(this.users.password.length>8){ 
+        Swal.fire({
+          position: "top-end",
+          icon: "success",
+          title: "Envio Registrado",
+          showConfirmButton: false,
+          timer: 1500
+        });
+        this.usuariosRegistrados.push(this.users);
 
-    // Limpiar los campos después de guardar los datos
-    this.users = {
-      email: '',
-      username: '',
-      password: ''
-    };
+        // Guardar la lista actualizada en el local storage
+        localStorage.setItem('users', JSON.stringify(this.usuariosRegistrados));
+    
+        // Limpiar los campos después de guardar los datos
+        this.users = {
+          email: '',
+          username: '', 
+          password: ''
+        };
+      
+
+      }else{
+        Swal.fire({
+          icon: "error",
+          title: "Oops...",
+          text: "Ingrese un Una contraseña de minimo 8 caracteres !",
+        });
+      }
+    }
+    else{
+      Swal.fire({
+        icon: "error",
+        title: "Oops...",
+        text: "Ingrese un Correo Valido!",
+      });
+    }
+  }
+   
   }
 
   click() {
-    console.log("¡Hola!");
     this.router.navigateByUrl("/");
   }
 }
